整理部分代码，请记得随时清理编译缓存，有些莫名其妙的垃圾Bug来自缓存

+++ src/main/java/otakuplus/straybird/othellogameserver/models/UserOnline.java
@@ -3,43 +3,43 @@ package otakuplus.straybird.othellogameserver.models;
 import javax.persistence.*;
 
 @Entity
-@Table(name="OthelloUserOnline")
+@Table(name = "OthelloUserOnline")
 public class UserOnline {
-	public static final int ONLINE = 100;
-	public static final int STAND_BY = 101;
-	public static final int GAMING = 102;
-	public static final int OFFLINE= 199;
+    public static final int ONLINE = 100;
+    public static final int STAND_BY = 101;
+    public static final int GAMING = 102;
+    public static final int OFFLINE = 199;
 
-	@Id
-	private Long userId;
-	private int onlineState;
+    @Id
+    private Integer userId;
+    private int onlineState;
 
-	@MapsId
-	@OneToOne(cascade=CascadeType.ALL)
-	@JoinColumn(name="userId")
-	private User user;
+    @MapsId
+    @OneToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "userId")
+    private User user;
 
-	public Long getUserId() {
-		return userId;
-	}
+    public Integer getUserId() {
+        return userId;
+    }
 
-	public void setUserId(Long userId) {
-		this.userId = userId;
-	}
+    public void setUserId(Integer userId) {
+        this.userId = userId;
+    }
 
-	public int getOnlineState() {
-		return onlineState;
-	}
+    public int getOnlineState() {
+        return onlineState;
+    }
 
-	public void setOnlineState(int onlineState) {
-		this.onlineState = onlineState;
-	}
+    public void setOnlineState(int onlineState) {
+        this.onlineState = onlineState;
+    }
 
-	public User getUser(){
-		return this.user;
-	}
+    public User getUser() {
+        return this.user;
+    }
 
-	public void setUser(User user){
-		this.user = user;
-	}
+    public void setUser(User user) {
+        this.user = user;
+    }
 }
